Made repulsive spline feed inherit from the Feed class and move it into the feed module. Unittests were modified to accommodate for the changes.
